Update C program file: 100-atoi.c

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -10,7 +10,7 @@ int _atoi(char *s)
 {
 	int i;
 	int sign;
-	int num;
+	unsigned int num;
 
 	i = 0;
 	sign = 1;
@@ -24,7 +24,7 @@ int _atoi(char *s)
 
 		for (; s[i] >= '0' && s[i] <= '9'; i++)
 		{
-			num = num * 10 + s[i] - '0';
+			num = num * 10 + ((int)s[i] - '0');
 		}
 
 		if (num > 0 || s[i] == '\0')
